backend/requirements.txt
google-genai>=1.14.0
python-dotenv==1.2.2
